+++ ErsatzTV.Core/Interfaces/Emby/IEmbyPathReplacementService.cs
@@ -6,4 +6,5 @@ public interface IEmbyPathReplacementService
 {
     Task<string> GetReplacementEmbyPath(int libraryPathId, string path, bool log = true);
     string GetReplacementEmbyPath(List<EmbyPathReplacement> pathReplacements, string path, bool log = true);
+    string ReplaceNetworkPath(EmbyMediaSource embyMediaSource, string path, string networkPath, string replacement);
 }