+++ ErsatzTV.Infrastructure/Data/Configurations/Library/EmbyLibraryConfiguration.cs
@@ -6,6 +6,12 @@ namespace ErsatzTV.Infrastructure.Data.Configurations;
 
 public class EmbyLibraryConfiguration : IEntityTypeConfiguration<EmbyLibrary>
 {
-    public void Configure(EntityTypeBuilder<EmbyLibrary> builder) =>
+    public void Configure(EntityTypeBuilder<EmbyLibrary> builder)
+    {
         builder.ToTable("EmbyLibrary");
+
+        builder.HasMany(l => l.PathInfos)
+            .WithOne()
+            .OnDelete(DeleteBehavior.Cascade);
+    }
 }